Move javassist helper methods into own class
